Adjust meeting flow test
