+++ cypress/integration/meetings-flow.spec.ts
@@ -1,92 +1,46 @@
 import { loginAs } from "../login"
+import { createMeeting } from "../meeting"
+import { createDefaultSchedule } from "../schedule"
 import { url } from "../support/url"
 import * as uuid from "uuid"
 import { addDays, format, isFriday, isSaturday } from "date-fns"
 
 import { johnDoe } from "../../db/seed-data"
-import { createDefaultSchedule } from "./schedules.spec"
 
 describe("Meetings Flow", () => {
-  const link = uuid.v4()
-  const meetingName = `My Test Meeting-${Math.floor(Math.random() * 10000)}`
-  const date = new Date()
-  const dateToSelect = isFriday(date) || isSaturday(date) ? addDays(date, 3) : addDays(date, 1)
-  const dateStart = format(date, "dd.MM.y")
-  const dateEnd = format(addDays(date, 14), "dd.MM.y")
-  it("Enter General Information", () => {
+  it("Create Meeting", () => {
+    const meetingLink = uuid.v4()
     loginAs(johnDoe)
-
-    cy.visit(url("/meetings"))
-
-    cy.contains("Create new Meeting").click()
-    cy.get("#name").type(meetingName)
-    cy.get("#link").type(link)
-    cy.get("#description").type("Lorem ipsum, dolor sit amet.")
-    cy.get("#location").type("Berlin")
-
-    cy.get("#submit").click()
-
-    cy.get("#duration-30").click()
-
-    cy.get("#range-start").type(dateStart)
-    cy.get("#range-end").type(dateEnd)
-  })
-  it("Creates Schedule from Meeting Modal", () => {
-    const scheduleName = uuid.v4()
-
-    createDefaultSchedule(scheduleName)
-
-    cy.get("#select-schedule").select(scheduleName)
-
-    cy.get("#submit").click()
-  })
-
-  it("Has invitee page", () => {
-    cy.contains(url(`/schedule/${johnDoe.username}/${link}`))
-    cy.visit(url(`/schedule/${johnDoe.username}/${link}`))
-  })
-
-  it("Selects Slot", () => {
-    cy.contains(format(dateToSelect, "d")).click()
-    cy.contains("button", /10:30.*-11:00.*/).click()
-
-    cy.contains("Schedule!").click()
+    createMeeting(uuid.v4(), meetingLink)
+    cy.contains(url(`/schedule/${johnDoe.username}/${meetingLink}`))
   })
 
   it("Creates Booking", () => {
+    loginAs(johnDoe)
+    const meetingLink = uuid.v4()
+    const meetingName = uuid.v4()
+    createMeeting(meetingName, meetingLink)
+    cy.visit(url(`/schedule/${johnDoe.username}/${meetingLink}`))
+    cy.wait(4000)
+    const date = new Date()
+    const dateToSelect = isFriday(date) || isSaturday(date) ? addDays(date, 3) : addDays(date, 1)
+    cy.contains(format(dateToSelect, "d")).click()
+    cy.wait(2000)
+    cy.contains("button", /11:30.*-12:00.*/).click()
+    cy.contains("Schedule!").click()
     cy.get("#formBasicEmail").type("test-receiver@kalle.app")
-
     cy.contains("Submit!").click()
-
     cy.wait(5000)
-
     cy.request("http://localhost:8025/api/v2/messages").then((response) => {
       const {
         items: [newestMail],
       } = response.body
-
       cy.log(response.body)
-
       expect(newestMail).to.exist
-
       expect(newestMail.Content.Headers.Subject[0]).to.equal(
-        `New appointment: ${meetingName} - 10:30, ${format(dateToSelect, "dd.MM.y")} with john.doe`
+        `New appointment: ${meetingName} - 11:30, ${format(dateToSelect, "dd.MM.y")} with john.doe`
       )
       expect(newestMail.Content.Headers.To[0]).to.equal("test-receiver@kalle.app")
     })
-  })
-  it("Creates Meeting in Overview", () => {
-    cy.visit(url(`/meetings`))
-    cy.contains(meetingName)
-
-    cy.contains(meetingName).parent().parent().contains("Delete").click()
-    cy.contains(meetingName).parent().parent().parent().contains("There are still active bookings")
-
-    cy.contains(meetingName).parent().parent().parent().contains("Copy Link")
-
-    cy.get(`button[id="booking-btn-${meetingName}"]`).click()
-    cy.contains("Lorem ipsum, dolor sit amet.")
-    cy.contains(format(dateToSelect, "d"))
-    cy.contains("test-receiver@kalle.app")
   })
 })
